JSを分割、srcフォルダにまとめる
